Uploading Paper
Traditional segmentations rely on proxy and claimed information when identifying media targeting opportunities. Mindshare has partnered with a 3rd party data supplier to augment survey data with behavioural and targetable 3rd party segment information so that the targeting information is based upon actual behaviour as opposed to claimed behaviour. This data augmentation allows for the creation of segmentations with direct buyable components to be created. With information around 3rd party segment membership the translation into a buyable strategy becomes more streamlined as well as providing a level of information prior to a campaign that can have an impact upon optimisation strategies due to a higher level of knowledge surrounding websites visited and potential audience overlap.